Added textures (mob and towers) Now in preparation screen able to fast move items by double clicking (store has this feature too) Added new spells and mob abilities Steam machine (and his support) was added (aura tower)

+++ core/src/com/darkhouse/gdefence/Objects/SpellObject.java
@@ -57,7 +57,7 @@ public abstract class SpellObject extends Ability.AbilityPrototype implements Ex
 
     }
 
-    public void addExp(int value){
+    public void addExp(float value){
         totalExp += value;
         updateExp();
     }
